fix: extracted reusable steps navigation component
